pyproject.toml
[project]
name = "ecdsattack"
version = "0.2.0"
description = "ecdsattack is a package which retrieves private keys from faulted ECDSA signatures."
license = {file = "LICENSE"}
readme = "README.md"
maintainers = [
  {name = "Adrian Thillard", email = "adrian.thillard@ledger.fr"},
  {name = "Jean-Baptiste Bédrune", email = "jean-baptiste.bedrune@ledger.fr"}
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python"
]

dependencies = [
  "ecdsa >= 0.17.0",
  "python-bitcoinlib >= 0.11.0", # For Bitcoin blockchain interaction
  "pymongo >= 4.0.0",  # For MongoDB database
  "motor >= 3.0.0", # For async MongoDB
  "pydantic >= 2.0.0", # For data validation
  "numpy >= 1.21.0",  # For numerical computations
  "scipy >= 1.7.0",  # For scientific computing
  "fpylll >= 0.5.0",  # For lattice reduction
  "PyYAML >= 6.0", # For YAML configuration parsing
  "requests >= 2.26.0",  # For HTTP requests
  "python-dotenv >= 0.19.0",  # For environment variables
  "asyncio >= 3.4.3",  # For async operations
  "pandas >= 1.3.0",  # For data manipulation
]

[project.optional-dependencies]
test = [
  "pytest >= 6.0.0",
  "pytest-asyncio >= 0.18.0",
  "pytest-cov >= 3.0.0",
  "black >= 22.0.0",  # For code formatting
  "isort >= 5.10.0",  # For import sorting
  "mypy >= 0.910",  # For type checking
  "flake8 >= 4.0.0"  # For linting
]

[build-system]
requires = [
    "setuptools>=40.6.0",
    "tomli"
]
build-backend = "setuptools.build_meta"
